# dash-backend/rest_api/swagger/index.yaml
definitions:
  Address:
    properties:
      city:
        $ref: '#/definitions/City'
      city_id:
        format: int32
        type: integer
      country:
        $ref: '#/definitions/Country'
      country_id:
        format: int32
        type: integer
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      raw:
        type: string
      state:
        $ref: '#/definitions/State'
      state_id:
        format: int32
        type: integer
      zip_code:
        format: int32
        type: integer
    type: object
  City:
    properties:
      country:
        $ref: '#/definitions/Country'
      country_id:
        format: int32
        type: integer
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      lat:
        format: float
        type: number
      long:
        format: float
        type: number
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      state:
        $ref: '#/definitions/State'
      state_id:
        format: int32
        type: integer
    type: object
  Country:
    properties:
      code:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    type: object
  DealerRequest:
    properties:
      address:
        items:
          $ref: '#/definitions/Address'
        type: array
      creation_date:
        format: date-time
        readOnly: true
        type: string
      credit_information:
        type: object
      dealer_id:
        format: int32
        type: integer
        x-nullable: true
      dealer_option:
        type: object
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      pricing_detail:
        type: object
      requester_email:
        type: string
      requester_phone_number:
        type: string
      status:
        enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        type: string
      user:
        $ref: '#/definitions/User'
      user_id:
        format: int32
        type: integer
      vehicle:
        $ref: '#/definitions/VehicleSchema'
      vehicle_id:
        format: int32
        type: integer
    required:
    - vehicle_id
    type: object
  DealerSchedulerSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      dealer_id:
        format: int32
        type: integer
        x-nullable: true
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      schedule_note:
        type: string
      schedule_time:
        format: date-time
        type: string
      schedule_title:
        type: string
      user:
        $ref: '#/definitions/User'
      user_id:
        format: int32
        type: integer
      vehicle:
        $ref: '#/definitions/VehicleSchema'
      vehicle_id:
        format: int32
        type: integer
    type: object
  DealerSchema:
    properties:
      address:
        items:
          $ref: '#/definitions/Address'
        type: array
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      lat:
        format: float
        type: number
      long:
        format: float
        type: number
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    required:
    - name
    type: object
...
  PriceOptionAssignmentSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      end_date:
        format: date-time
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      region_id:
        format: int32
        type: integer
      start_date:
        format: date-time
        type: string
      value_in_cent:
        format: int32
        type: integer
      vehicle_option_item_id:
        format: int32
        type: integer
    type: object
  PriceVehicleAssignmentSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      end_date:
        format: date-time
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      region_id:
        format: int32
        type: integer
      start_date:
        format: date-time
        type: string
      value_in_cent:
        format: int32
        type: integer
      vehicle_id:
        format: int32
        type: integer
    type: object
  PriceVehicleModelStyleAssignmentSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      end_date:
        format: date-time
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      region_id:
        format: int32
        type: integer
      start_date:
        format: date-time
        type: string
      value_in_cent:
        format: int32
        type: integer
      vehicle_model_style_id:
        format: int32
        type: integer
    type: object
  Region:
    properties:
      code:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    type: object
  State:
    properties:
      code:
        type: string
      country:
        $ref: '#/definitions/Country'
      country_id:
        format: int32
        type: integer
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      region:
        $ref: '#/definitions/Region'
      region_id:
        format: int32
        type: integer
    type: object
  User:
    properties:
      address:
        type: string
      avatar_url:
        type: string
      email:
        format: email
        type: string
      facebook_id:
        type: string
      firstName:
        type: string
      google_id:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      lastName:
        type: string
      password:
        type: string
      phone_number:
        type: string
      title:
        type: string
      type:
        enum:
        - HOST
        - TENANT
        - EXTERNAL
        type: string
    required:
    - email
    - firstName
    - lastName
    type: object
...
  VehicleBody:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  VehicleBrand:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  VehicleDriveTrain:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  VehicleEngine:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      horse_power:
        format: int32
        type: integer
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      label:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      torque:
        format: int32
        type: integer
      type:
        enum:
        - OTHERS
        - GAS
        - DIESEL
        - ELECTRIC
        type: string
    required:
    - type
    type: object
  VehicleGalleryItemSchema:
    properties:
      asset_path:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      order:
        format: int32
        type: integer
      type:
        enum:
        - COVER
        - INTERIOR
        - INTERIOR_COLOR
        - EXTERIOR
        - EXTERIOR_COLOR
        - BRAND_LOGO
        - MODEL_AVATAR
        - YEAR_MODEL_AVATAR
        - VIDEO_INTRODUCTION
        - MODEL_STYLE_AVATAR
        - OPTION_ITEM_AVATAR
        - VEHICLE_AVATAR
        - OTHERS
        type: string
    required:
    - asset_path
    - type
    type: object
...
  VehicleInfo:
    properties:
      city_mpg:
        format: int32
        type: integer
      combined_mpg:
        format: int32
        type: integer
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      doors:
        format: int32
        type: integer
      highway_mpg:
        format: int32
        type: integer
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      safety_rate:
        format: int32
        type: integer
      seats:
        format: int32
        type: integer
      vehicle_warranty:
        $ref: '#/definitions/VehicleWarranty'
    type: object
  VehicleModel:
    properties:
      description:
        type: string
      name:
        type: string
      vehicle_brand:
        allOf:
        - $ref: '#/definitions/VehicleBrand'
        readOnly: true
    required:
    - description
    - name
    type: object
  VehicleModel1:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  VehicleModelStyle:
    properties:
      base_MSRP:
        format: float
        type: number
      body:
        allOf:
        - $ref: '#/definitions/VehicleBody'
        readOnly: true
      body_id:
        format: int32
        type: integer
      consumer_info:
        type: object
      creation_date:
        format: date-time
        readOnly: true
        type: string
      current_price_assigment:
        items:
          $ref: '#/definitions/PriceVehicleModelStyleAssignmentSchema'
        type: array
      description:
        type: string
      drive_trains:
        items:
          $ref: '#/definitions/VehicleDriveTrain'
        type: array
      engines:
        items:
          $ref: '#/definitions/VehicleEngine'
        type: array
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      image_url:
        type: string
      include_feature_description:
        type: string
      infos:
        items:
          $ref: '#/definitions/VehicleInfo'
        type: array
      manufacture_code:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/VehicleOptionItem'
        type: array
      pricing_options:
        type: object
      standard_feature:
        type: object
      style_features:
        items:
          type: string
        type: array
      technical_feature:
        type: object
      transmissions:
        items:
          $ref: '#/definitions/VehicleTransmission'
        type: array
      vehicle_year_model:
        allOf:
        - $ref: '#/definitions/VehicleYearModel'
        readOnly: true
      vehicle_year_model_id:
        format: int32
        type: integer
      vif_num:
        format: int32
        readOnly: true
        type: integer
    required:
    - description
    - name
    type: object
...
  VehicleOptionItem:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      current_price_assigment:
        items:
          $ref: '#/definitions/PriceOptionAssignmentSchema'
        type: array
      description:
        type: string
      detail:
        type: object
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      is_popular:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      type:
        enum:
        - INTERIOR_COLOR
        - EXTERIOR_COLOR
        - PACKAGE
        - OPTION
        - OTHERS
        type: string
    required:
    - description
    - type
    type: object
...
  VehicleSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      current_owner:
        $ref: '#/definitions/User'
      current_owner_id:
        format: int32
        type: integer
      current_price_assigment:
        $ref: '#/definitions/PriceVehicleAssignmentSchema'
      custom_options:
        items:
          $ref: '#/definitions/VehicleOptionItem'
        type: array
      description:
        type: string
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      vehicle_brand:
        allOf:
        - $ref: '#/definitions/VehicleBrand'
        readOnly: true
      vehicle_brand_id:
        format: int32
        type: integer
      vehicle_model:
        allOf:
        - $ref: '#/definitions/VehicleModel1'
        readOnly: true
      vehicle_model_id:
        format: int32
        type: integer
      vehicle_model_style:
        allOf:
        - $ref: '#/definitions/VehicleModelStyle'
        readOnly: true
      vehicle_model_style_id:
        format: int32
        type: integer
    required:
    - description
    - name
    type: object
...
  VehicleTransmission:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      label:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      speeds:
        format: int32
        type: integer
      type:
        enum:
        - OTHERS
        - AUTOMATIC
        - MANUAL
        type: string
    required:
    - type
    type: object
  VehicleWarranty:
    properties:
      anti_corrosion:
        type: string
      comprehensive:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      paint:
        type: string
      power_train:
        type: string
      roadside_assistance:
        type: string
    type: object
  VehicleYearModel:
    properties:
      description:
        type: string
      name:
        type: string
      vehicle_model:
        allOf:
        - $ref: '#/definitions/VehicleModel'
        readOnly: true
    required:
    - description
    - name
    type: object
...
info:
  title: Swagger Petstore
  version: 1.0.0
paths:
...
  /dealer_requests:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: DealerRequestSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: DealerBody
        required: 'true'
        schema:
          $ref: '#/definitions/DealerRequest'
      responses:
        '200':
          content:
            application/json:
              schema: DealerRequestSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /dealer_requests/{dealer_request_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: DealerRequestSchema
        '400': &id001
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id002
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id003
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: DealerRequestSchema
        '400': *id001
        '401': *id002
        5XX: *id003
    parameters:
    - description: DealerRequestId
      in: path
      name: dealer_request_id
      required: true
    put:
      parameters:
      - in: body
        name: DealerRequestBody
        required: 'true'
        schema:
          $ref: '#/definitions/DealerRequest'
      responses:
        '200':
          content:
            application/json:
              schema: DealerRequestSchema
        '400': *id001
        '401': *id002
        5XX: *id003
  /dealer_schedulers:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchedulerSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: DealerBody
        required: 'true'
        schema:
          $ref: '#/definitions/DealerSchedulerSchema'
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchedulerSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /dealer_schedulers/{dealer_scheduler_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchedulerSchema
        '400': &id004
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id005
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id006
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchedulerSchema
        '400': *id004
        '401': *id005
        5XX: *id006
    parameters:
    - description: DealerSchedulerId
      in: path
      name: dealer_scheduler_id
      required: true
    put:
      parameters:
      - in: body
        name: DealerSchedulerBody
        required: 'true'
        schema:
          $ref: '#/definitions/DealerSchedulerSchema'
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchedulerSchema
        '400': *id004
        '401': *id005
        5XX: *id006
  /dealer_schedulers/{dealer_scheduler_id}/month/{month_num}:
    get:
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchedulerSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    parameters:
    - description: DealerSchedulerId
      in: path
      name: dealer_scheduler_id
      required: true
    - description: Number Of Month
      in: path
      name: month_num
      required: true
  /dealers:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: DealerBody
        required: 'true'
        schema:
          $ref: '#/definitions/DealerSchema'
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /dealers/{dealer_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchema
        '400': &id007
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id008
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id009
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchema
        '400': *id007
        '401': *id008
        5XX: *id009
    parameters:
    - description: DealerId
      in: path
      name: dealer_id
      required: true
    put:
      parameters:
      - in: body
        name: DealerBody
        required: 'true'
        schema:
          $ref: '#/definitions/DealerSchema'
      responses:
        '200':
          content:
            application/json:
              schema: DealerSchema
        '400': *id007
        '401': *id008
        5XX: *id009
...
swagger: 2.0.0